refactor(test): consolidate introspection unit tests
Merge tests/unit/test_parameter_aliases.py, tests/unit/test_handler_signature.py, tests/unit/test_descriptions.py, and tests/unit/test_description_rendering_endpoints.py into tests/unit/test_introspection.py to reduce test file sprawl. Clean up unused imports and group tests using classes. Delete the redundant files.

TAG=agy
CONV=4db20f37-0e22-4128-983c-20e382b47363
